chore: release v0.0.11

+++ server/src/server.ts
@@ -18,6 +18,9 @@ import { fork } from 'child_process';
 import { capture } from "./utils/analytics";
 import swaggerUi from 'swagger-ui-express';
 import swaggerSpec from './swagger/config';
+
+import session from 'express-session';
+
 import Run from './models/Run';
 
 const app = express();
@@ -26,6 +29,16 @@ app.use(cors({
   credentials: true,
 }));
 app.use(express.json());
+
+
+app.use(
+  session({
+    secret: 'your_secret_key', // Replace with a secure secret key
+    resave: false, // Do not resave the session if it hasn't changed
+    saveUninitialized: true, // Save new sessions
+    cookie: { secure: false }, // Set to true if using HTTPS
+  })
+);
 
 const server = http.createServer(app);
 
@@ -65,8 +78,11 @@ readdirSync(path.join(__dirname, 'api')).forEach((r) => {
 
 const isProduction = process.env.NODE_ENV === 'production';
 const workerPath = path.resolve(__dirname, isProduction ? './worker.js' : './worker.ts');
+const recordingWorkerPath = path.resolve(__dirname, isProduction ? './pgboss-worker.js' : './pgboss-worker.ts');
 
 let workerProcess: any;
+let recordingWorkerProcess: any;
+
 if (!isProduction) {
   workerProcess = fork(workerPath, [], {
     execArgv: ['--inspect=5859'],
@@ -79,6 +95,19 @@ if (!isProduction) {
   });
   workerProcess.on('exit', (code: any) => {
     console.log(`Worker exited with code: ${code}`);
+  });
+
+  recordingWorkerProcess = fork(recordingWorkerPath, [], {
+    execArgv: ['--inspect=5860'],
+  });
+  recordingWorkerProcess.on('message', (message: any) => {
+    console.log(`Message from recording worker: ${message}`);
+  });
+  recordingWorkerProcess.on('error', (error: any) => {
+    console.error(`Error in recording worker: ${error}`);
+  });
+  recordingWorkerProcess.on('exit', (code: any) => {
+    console.log(`Recording worker exited with code: ${code}`);
   });
 }
 
@@ -132,7 +161,8 @@ process.on('SIGINT', async () => {
   }
   
   if (!isProduction) {
-    workerProcess.kill();
+    if (workerProcess) workerProcess.kill();
+    if (recordingWorkerProcess) recordingWorkerProcess.kill();
   }
   process.exit();
 });
